+++ Controllers/Views/IEngineTestConsoleView.cs
@@ -9,7 +9,7 @@ namespace Presentation.Views
 {
     public interface IEngineTestConsoleView : IConsoleView
     {
-        event Action<int> TemperatureSelected;
+        event Action<double> TemperatureSelected;
         event Action<int> EngineKindSelected;
         event Action<int> TestTypeSelected;
         void InvokeTemperatureInput();
